--- app/cdap/components/ConnectionList/Components/SubHeader/styles.ts
@@ -1,47 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import { makeStyles } from '@material-ui/styles';
-import grey from '@material-ui/core/colors/grey';
-import { blue } from '@material-ui/core/colors';
-
-export const useStyles = makeStyles({
-  breadcrumbContainer: {
-    borderBottom: `1px solid ${grey[300]}`,
-    display: 'flex',
-    justifyContent: 'space-between',
-    height: '48px',
-    alignItems: 'center',
-    paddingRight: 30,
-    '& .MuiBreadcrumbs-root': {
-      fontSize: '14px',
-      fontWeight: '400',
-    },
-    '& .MuiTypography-root': {
-      color: '#000000',
-      lineHeight: '21px',
-    },
-    '& a': {
-      color: blue[500],
-      lineHeight: '21px',
-    },
-  },
-  selectPrevPage: {
-    display: 'flex',
-    gap: '8px',
-    alignItems: 'center',
-  },
-});
