+++ app/cdap/components/ConnectionList/Components/TabLabelCanBrowse/index.tsx
@@ -14,7 +14,6 @@
  * the License.
  */
 
-import { Box } from '@material-ui/core';
 import CustomTooltip from 'components/ConnectionList/Components/CustomTooltip';
 import TabLabelItem from 'components/ConnectionList/Components/LabelItemCanBrowse';
 import * as React from 'react';
@@ -50,9 +49,7 @@ export default function({ label, count, columnIndex, icon }: ITabLabelCanBrowseP
       key={`tooltip-${columnIndex}`}
       data-testid="connections-tab-ref-label-browse"
     >
-      <Box>
-        <TabLabelItem myLabelRef={myLabelRef} icon={icon} label={label} count={count} />
-      </Box>
+      <TabLabelItem myLabelRef={myLabelRef} icon={icon} label={label} count={count} />
     </CustomTooltip>
   ) : (
     <TabLabelItem
